+++ R/enspls.fs.R
@@ -5,7 +5,9 @@
 #' @param x Predictor matrix.
 #' @param y Response vector.
 #' @param maxcomp Maximum number of components included within each model.
-#' If not specified, will use 5 by default.
+#' If not specified, will use \code{5} by default.
+#' @param cvfolds Number of cross-validation folds used in each model
+#' for automatic parameter selection, default is \code{5}.
 #' @param alpha Parameter (grid) controlling sparsity of the model.
 #' If not specified, default is \code{seq(0.2, 0.8, 0.2)}.
 #' @param reptimes Number of models to build with Monte-Carlo resampling
@@ -46,10 +48,12 @@
 #' plot(fs, type = 'boxplot', limits = c(0.05, 0.95), nvar = 10)
 
 enspls.fs = function(x, y,
-                     maxcomp = 5L,
-                     alpha = seq(0.2, 0.8, 0.2),
+                     maxcomp  = 5L,
+                     cvfolds  = 5L,
+                     alpha    = seq(0.2, 0.8, 0.2),
                      reptimes = 500L,
-                     method = c('mc', 'boot'), ratio = 0.8,
+                     method   = c('mc', 'boot'),
+                     ratio    = 0.8,
                      parallel = 1L) {
 
   if (missing(x) | missing(y)) stop('Please specify both x and y')
@@ -74,7 +78,7 @@ enspls.fs = function(x, y,
       xtmp = x[samp.idx[[i]], ]
       xtmp = scale(xtmp, center = TRUE, scale = TRUE)
       ytmp = y[samp.idx[[i]]]
-      coeflist[[i]] = enspls.fs.core(xtmp, ytmp, maxcomp, alpha)
+      coeflist[[i]] = enspls.fs.core(xtmp, ytmp, maxcomp, cvfolds, alpha)
     }
 
   } else {
@@ -84,7 +88,7 @@ enspls.fs = function(x, y,
       xtmp = x[samp.idx[[i]], ]
       xtmp = scale(xtmp, center = TRUE, scale = TRUE)
       ytmp = y[samp.idx[[i]]]
-      enspls.fs.core(xtmp, ytmp, maxcomp, alpha)
+      enspls.fs.core(xtmp, ytmp, maxcomp, cvfolds, alpha)
     }
 
   }
@@ -115,23 +119,29 @@ enspls.fs = function(x, y,
 #'
 #' @keywords internal
 
-enspls.fs.core = function(xtmp, ytmp, maxcomp, alpha) {
+enspls.fs.core = function(xtmp, ytmp, maxcomp, cvfolds, alpha) {
 
   invisible(
     capture.output(
-      spls.cvfit <- cv.spls(xtmp, ytmp,
-                            fold = 5,
-                            K = maxcomp, eta = alpha,
-                            scale.x = TRUE, scale.y = FALSE,
+      spls.cvfit <- cv.spls(xtmp,
+                            ytmp,
+                            fold    = cvfolds,
+                            K       = maxcomp,
+                            eta     = alpha,
+                            scale.x = TRUE,
+                            scale.y = FALSE,
                             plot.it = FALSE)))
 
   # select best component number and alpha using adjusted CV
   cv.bestcomp  = spls.cvfit$'K.opt'
   cv.bestalpha = spls.cvfit$'eta.opt'
 
-  spls.fit = spls(xtmp, ytmp,
-                  K = cv.bestcomp, eta = cv.bestalpha,
-                  scale.x = TRUE, scale.y = FALSE)
+  spls.fit = spls(xtmp,
+                  ytmp,
+                  K       = cv.bestcomp,
+                  eta     = cv.bestalpha,
+                  scale.x = TRUE,
+                  scale.y = FALSE)
 
   spls.coef = coef(spls.fit)
   spls.coef.vec = as.vector(spls.coef)